Almost done
menší změny, fixnutí pagingu, dodělání administrace, úprava bootstrapu, css

+++ src/main/java/com/example/rocnikprojekt/DeleteCharacterServlet.java
@@ -10,13 +10,21 @@ import jakarta.servlet.annotation.*;
 public class DeleteCharacterServlet extends HttpServlet {
 
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        if(request.getParameter("id") != null && !request.getParameter("id").isEmpty()){
-            int id = Integer.parseInt(request.getParameter("id"));
-            CharacterDAO characterDAO = new CharacterDAO();
-            characterDAO.deleteCharacter(id);
+        //if session doesnt exist, dont create new one.
+        HttpSession session = request.getSession(false);
+        //checking if user is logged in - otherwise - bye bye
+        if (session != null && session.getAttribute("user") != null) {
 
+            if (request.getParameter("id") != null && !request.getParameter("id").isEmpty()) {
+                int id = Integer.parseInt(request.getParameter("id"));
+                CharacterDAO characterDAO = new CharacterDAO();
+                characterDAO.deleteCharacter(id);
+
+            }
+            response.sendRedirect("/approve?CharacterCount=" + request.getParameter("CharacterCount"));
+        }else{
+            response.sendRedirect("login.jsp");
         }
-        response.sendRedirect("/approve?CharacterCount=" + request.getParameter("CharacterCount"));
     }
 
     public void destroy() {
